move password manager to content settings
